ref(lang): rename BigInteger to BigInt for Clojure-aligned naming
Step 4/6 of the class-rename recipe in #2000: align the Phel runtime
class name with the `BigInt` auto-refer alias added in #1996.

- `Phel\Lang\BigInteger` -> `Phel\Lang\BigInt`
- Test file renamed accordingly
- Default alias map updated: `'BigInt' => Phel\Lang\BigInt::class`
- `core/predicates.phel`, `core/math.phel`, `core/booleans.phel` switch
  from `(:use Phel.Lang.BigInteger)` to `(:use Phel.Lang.BigInt)`
- Docs refreshed: `numeric-tower.md`, `clojure-migration.md`,
  `patterns.md`, example phel files

BC break: any caller importing `\Phel\Lang\BigInteger` or writing
`(:use Phel.Lang.BigInteger)` in a Phel ns form must switch to `BigInt`
(or drop the `(:use ...)` line, as `BigInt` is auto-referred since #1996).

Related to #2000

+++ src/php/Compiler/Domain/Analyzer/Environment/SymbolResolver.php
@@ -30,7 +30,7 @@ final readonly class SymbolResolver
      * up by the auto-resolution fallback below and need no entry here.
      */
     private const array CLOJURE_CLASS_RENAMES = [
-        'BigInt' => 'BigInteger',
+        'BigInt' => 'BigInt',
         'Ratio' => 'Rational',
     ];
 
@@ -202,7 +202,7 @@ final readonly class SymbolResolver
      * Maps `clojure.lang.X` references to the matching Phel runtime class.
      * Auto-resolves when `\Phel\Lang\X` exists (covers `Symbol`, `Keyword`,
      * `Atom`, etc.); a small rename table handles cases where Phel
-     * diverged from the Clojure name (`BigInt` vs `BigInteger`, `Ratio` vs
+     * diverged from the Clojure name (`BigInt` vs `BigInt`, `Ratio` vs
      * `Rational`). Returns null when no Phel counterpart exists.
      */
     private function remapClojureClassFqn(string $name): ?string
