+++ tests/php/Unit/Compiler/Emitter/OutputEmitter/LiteralEmitterTest.php
@@ -6,7 +6,7 @@ namespace PhelTest\Unit\Compiler\Emitter\OutputEmitter;
 
 use Phel\Compiler\CompilerFactory;
 use Phel\Compiler\Domain\Emitter\OutputEmitter\LiteralEmitter;
-use Phel\Lang\BigInteger;
+use Phel\Lang\BigInt;
 use Phel\Lang\Rational;
 use Phel\Shared\Printer\Printer;
 use PHPUnit\Framework\TestCase;
@@ -28,44 +28,44 @@ final class LiteralEmitterTest extends TestCase
 
     public function test_emit_big_integer_uses_from_string_factory(): void
     {
-        $this->literalEmitter->emitLiteral(BigInteger::fromString('123456789012345678901234567890'));
+        $this->literalEmitter->emitLiteral(BigInt::fromString('123456789012345678901234567890'));
 
-        $this->expectOutputString('\\' . BigInteger::class . '::fromString("123456789012345678901234567890")');
+        $this->expectOutputString('\\' . BigInt::class . '::fromString("123456789012345678901234567890")');
     }
 
     public function test_emit_negative_big_integer(): void
     {
-        $this->literalEmitter->emitLiteral(BigInteger::fromString('-99999999999999999999'));
+        $this->literalEmitter->emitLiteral(BigInt::fromString('-99999999999999999999'));
 
-        $this->expectOutputString('\\' . BigInteger::class . '::fromString("-99999999999999999999")');
+        $this->expectOutputString('\\' . BigInt::class . '::fromString("-99999999999999999999")');
     }
 
     public function test_emit_rational_uses_create_factory_with_big_integer_arguments(): void
     {
-        $rational = Rational::create(BigInteger::fromString('1'), BigInteger::fromString('2'));
+        $rational = Rational::create(BigInt::fromString('1'), BigInt::fromString('2'));
         self::assertInstanceOf(Rational::class, $rational);
 
         $this->literalEmitter->emitLiteral($rational);
 
         $this->expectOutputString(
             '\\' . Rational::class . '::create('
-            . '\\' . BigInteger::class . '::fromString("1"), '
-            . '\\' . BigInteger::class . '::fromString("2")'
+            . '\\' . BigInt::class . '::fromString("1"), '
+            . '\\' . BigInt::class . '::fromString("2")'
             . ')',
         );
     }
 
     public function test_emit_negative_rational(): void
     {
-        $rational = Rational::create(BigInteger::fromString('-3'), BigInteger::fromString('4'));
+        $rational = Rational::create(BigInt::fromString('-3'), BigInt::fromString('4'));
         self::assertInstanceOf(Rational::class, $rational);
 
         $this->literalEmitter->emitLiteral($rational);
 
         $this->expectOutputString(
             '\\' . Rational::class . '::create('
-            . '\\' . BigInteger::class . '::fromString("-3"), '
-            . '\\' . BigInteger::class . '::fromString("4")'
+            . '\\' . BigInt::class . '::fromString("-3"), '
+            . '\\' . BigInt::class . '::fromString("4")'
             . ')',
         );
     }
